Let Tensor comm wait for CUDA ops on new devices in RPC response
closes #53287

When using CUDA RPC, if the user function uses a new device which
is not used by any Tensors in the request, current TensorPipeAgent
cannot properly synchronize communication and computation. This is
because the agent cannot predict which devices a user function will
use and is not aware of which stream on that device it should wait
for. Two naive solutions might be 1) create a context/stream for
all available devices, 2) always synchronize with current/default
stream on the new device. However, neither of these would work. The
first one would create CUDA context on a device even if the device
was not used in the RPC. As a result, it would introduce additional
CUDA memory overhead, or worse, lead to deadlock. The second one
will sync all background RPCs to the same current/default stream,
which would interfere with user logic on the user main thread and
contentions on that stream could lead to perfermance problem and
correctness might be hard to argue.

This PR is step 1 towards addressing the problem. We now require
users to synchronize their new device streams to one of the current
stream on existing devices. Then, the agent would let communications
on the new device to wait for streams on existing devices.

After this, we still need to solve the problem of CPU-request +
CUDA-reponse, where we don't have a existing stream at all. One
possibility might be letting users to specify streams for the
user function.

ghstack-source-id: 9ca1934cfb5859cb64cd5b92cbcd5124d5af0dd4
Pull Request resolved: https://github.com/pytorch/pytorch/pull/53742